enable ALL the warnings
and fix ALL of them.

+++ src/filesystem.cpp
@@ -13,14 +13,14 @@ void FileMmap::open(const char *path) {
 		::close(fd);
 		throw std::exception();
 	}
-	void *res = mmap(nullptr, stat.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
+	void *res = mmap(nullptr, (size_t)stat.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
 	if (res == MAP_FAILED) {
 		::close(fd);
 		throw std::exception();
 	}
 	::close(fd);
 	base = (const uint8_t *)res;
-	len = stat.st_size;
+	len = (size_t)stat.st_size;
 }
 
 }
